+++ pytoil/cli/pull.py
@@ -17,121 +17,100 @@ from pytoil.config import Config
 from pytoil.git import Git
 from pytoil.repo import Repo
 
-app = typer.Typer(name="pull", no_args_is_help=True)
+app = typer.Typer()
 
 
-# Callback for documentation only
-@app.callback()
-def pull() -> None:
+@app.command()
+def pull(
+    projects: List[str] = typer.Argument(
+        None,
+        help="Name of the project(s) to pull down.",
+        show_default=False,
+    ),
+    force: bool = typer.Option(
+        False,
+        "--force",
+        "-f",
+        help="Force pull without confirmation.",
+        show_default=False,
+    ),
+    all_: bool = typer.Option(
+        False,
+        "--all",
+        "-a",
+        help="Pull down all of your projects.",
+        show_default=False,
+    ),
+) -> None:
     """
     Pull down your remote projects.
 
     The pull command provides easy methods for pulling down remote projects.
 
-    It is effectively a nice wrapper around git clone but you don't have to worry
-    about urls or what your cwd is, pull will grab your remote projects by
-    name and clone them to your configured projects directory.
+    It is effectively a nice wrapper around git clone but you don't have to
+    worry about urls or what your cwd is, pull will grab your remote projects
+    by name and clone them to your configured projects directory.
 
-    You can also use pull to batch clone multiple repos, even all of them if you're
-    into that sorta thing.
+    You can also use pull to batch clone multiple repos, even all of them ("--all/-a")
+    if you're into that sorta thing.
 
-    Any remote project that already exists locally will be skipped and
-    none of your local projects are changed in any way.
+    Any remote project that already exists locally will be skipped and none of
+    your local projects are changed in any way. pytoil will only pull down
+    those projects that don't already exist locally.
 
-    It's very possible to accidentally clone a lot of repos when using pull
-    so you will be prompted for confirmation before pytoil does anything.
+    It's very possible to accidentally clone a lot of repos when using pull so
+    you will be prompted for confirmation before gotoil does anything.
 
-    The "--force/-f" flag can be used to override this confirmation prompt
-    if desired.
-    """
+    The "--force/-f" flag can be used to override this confirmation prompt if
+    desired.
 
+    Examples:
 
-@app.command(name="all")
-def all_(
-    force: bool = typer.Option(
-        False,
-        "--force",
-        "-f",
-        help="Force pull without confirmation.",
-    )
-) -> None:
-    """
-    Pull down all your remote projects.
+    $ pytoil pull project1 project2 project3
 
-    This will look at the projects you already have locally and
-    on GitHub, calculate the difference and then use git to clone
-    each one to your configured projects directory.
+    $ pytoil pull project1 project2 project3 --force
 
-    Any projects that already exist locally are completely skipped and
-    will not be altered in any way.
+    $ pytoil pull --all
 
-    You will be prompted for confirmation unless the "--force/-f" flag
-    is used.
+    $ pytoil pull --all --force
     """
     config = Config.from_file()
     utils.warn_if_no_api_creds(config)
 
     api = API(username=config.username, token=config.token)
 
-    try:
-        with msg.loading("Calculating difference..."):
-            local_projects = utils.get_local_projects(path=config.projects_dir)
-            remote_projects = api.get_repo_names()
-
-            diff = remote_projects.difference(local_projects)
-    except httpx.HTTPStatusError as err:
-        utils.handle_http_status_errors(error=err)
-    else:
-
-        pull_diff(
-            diff=diff, remotes=remote_projects, force=force, config=config, git=Git()
+    if not projects and not all_:
+        msg.warn(
+            "If not using the '--all' flag, you must specify projects to pull.",
+            exits=1,
+            spaced=True,
         )
 
-
-@app.command()
-def these(
-    projects: List[str] = typer.Argument(
-        ..., help="Name of the project(s) to pull down."
-    ),
-    force: bool = typer.Option(
-        False,
-        "--force",
-        "-f",
-        help="Force pull without confirmation.",
-    ),
-) -> None:
-    """
-    Pull down specified remote projects.
-
-    This will look at the projects you've specified, those you already have locally, and
-    those on GitHub, calculate the difference and then use git to clone
-    each one to your configured projects directory.
-
-    Any projects that already exist locally are completely skipped and
-    will not be altered in any way.
-
-    You will be prompted for confirmation unless the "--force/-f" flag
-    is used.
-    """
-    config = Config.from_file()
-    utils.warn_if_no_api_creds(config)
-
-    api = API(username=config.username, token=config.token)
-
     try:
+        not_found = False
         with msg.loading("Calculating difference..."):
             local_projects = utils.get_local_projects(path=config.projects_dir)
             remote_projects = api.get_repo_names()
+
+            specified_remotes = remote_projects if all_ else set(projects)
 
             for project in projects:
                 if project not in remote_projects:
-                    msg.warn(
-                        f"{project!r} not found on GitHub. Was it a typo?",
-                        spaced=True,
-                        exits=1,
-                    )
+                    not_found = True
+                    not_found_project = project
+                    break
+                break
 
-            diff = set(projects).difference(local_projects)
+        if not_found:
+            # If we don't break out of the with block
+            # the spinner will go forever
+            msg.warn(
+                f"{not_found_project!r} not found on GitHub. Was it a typo?",
+                spaced=True,
+                exits=1,
+            )
+
+        diff = specified_remotes.difference(local_projects)
     except httpx.HTTPStatusError as err:
         utils.handle_http_status_errors(error=err)
     else: