+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.6.dev
 
 * NEW: User can now initialise basic starter templates for multiple languages (currently support python, go and rust)
+* NEW: `pytoil remove` and `pytoil pull` changed to take projects as arguments, `remove|pull all` functionality is now in the `--all` option.
 
 * FIX: pytoil now supports the `python.defaultInterpreterPath` workspace setting rather than the soon to be deprecated `python.pythonPath`. See: https://github.com/microsoft/vscode-python/issues/12313